minor cleanup in player.cpp and control.cpp - add inactive code to handle durability of staves on block - dir2sdir is debug-only - make base type of Abilities, SpellPages and ClassIconTbl BYTE - fix caching logic in SetSpellTrans - make fields of control.cpp static if possible - rename pPanelButtons to pPanelButtonCels - rename pFlasks to pFlaskCels - rename pDurIcons to pDurIconCels - rename pTalkPnl to pTalkPnlCel - rename pGBoxBuff to pGoldDropCel - rename pChrPanel to pChrPanelCel - rename pChrButtons to pChrButtonCels

+++ Source/player.h
@@ -90,7 +90,7 @@ extern const int StrengthTbl[NUM_CLASSES];
 extern const int MagicTbl[NUM_CLASSES];
 extern const int DexterityTbl[NUM_CLASSES];
 extern const int VitalityTbl[NUM_CLASSES];
-extern const int Abilities[NUM_CLASSES];
+extern const BYTE Abilities[NUM_CLASSES];
 extern const char *const ClassStrTbl[NUM_CLASSES];
 extern const unsigned SkillExpLvlsTbl[MAXSPLLEVEL + 1];
